Move location stuff to new "locations" app
