+++ backend/kds_stroy/settings.py
@@ -25,9 +25,9 @@ INSTALLED_APPS = [
     "users.apps.UsersConfig",
     "news.apps.NewsConfig",
     "orders.apps.OrdersConfig",
+    "locations.apps.LocationsConfig",
     "main.apps.MainConfig",
     "ads_mailing.apps.AdsMailingConfig",
-    "locations.apps.LocationsConfig",
     "verify_email.apps.VerifyEmailConfig",
     "sass_processor",
     "django_ckeditor_5",
