+++ backend/kds_stroy/urls.py
@@ -8,17 +8,16 @@ from users.views import RegistrationView
 
 urlpatterns = [
     path("", MainView.as_view(), name="home"),
-    path("terms/", TermsView.as_view(), name="personal_info_terms"),
     path("admin/", admin.site.urls),
-    path("verification/", include("verify_email.urls")),
-    path("auth/registration/",
-         RegistrationView.as_view(),
-         name="registration"),
+    path("auth/registration/", RegistrationView.as_view(), name="registration"),
     path("auth/", include("django.contrib.auth.urls")),
-    path("profile/", include("users.urls")),
-    path("orders/", include("orders.urls")),
+    path("locations/", include("locations.urls")),
     path("news/", include("news.urls")),
+    path("orders/", include("orders.urls")),
+    path("profile/", include("users.urls")),
     path("subs/", include("ads_mailing.urls")),
+    path("terms/", TermsView.as_view(), name="personal_info_terms"),
+    path("verification/", include("verify_email.urls")),
     path("ckeditor5/", include('django_ckeditor_5.urls')),
 ]
