+++ backend/locations/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+
+from locations.views import LocationAutocompleteView
+
+app_name = "locations"
+
+urlpatterns = [
+    path('autocomplete/',
+         LocationAutocompleteView.as_view(),
+         name='autocomplete')
+]
